Add Frame AutoGeneration Toggle Command
Adding a new command, that allows autogeneration of frames to be toggled. By default, autogeneration is enabled, which causes frames to be added and removed automatically. However, in the case that a user doesn't prefer this behavior, they can now disable autogeneration of frames using this toggle command. To re-enable autogeneration, the toggle command can be used.

Commands added:
- /lb ag
- /lb autogeneration